modules/gpu-server/requirements.txt
# GPU Server Dependencies
# NOTE: Heavy deps (torch, vllm, tensorflow, flash-attn) are installed
# directly in the Dockerfile for better layer caching. This file contains
# the lightweight application dependencies only.

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# HTTP client (for vLLM communication)
httpx>=0.26.0

# Image processing
Pillow>=10.2.0
numpy>=1.26.0

# Facial recognition
deepface>=0.0.79
tf-keras>=2.18.0

# Logging
loguru>=0.7.2

# Configuration
pyyaml>=6.0.1

# Utilities
python-multipart>=0.0.6
